print all data ok

+++ Program.cs
@@ -4,19 +4,19 @@ using var db = new DB();
 Console.WriteLine($"Database path: {db.path}.");
 
 /* Create - user */
-Console.WriteLine("Inserting - user");
+Console.WriteLine("Inserting > user");
 db.Users.Add(new User { Name = "Yummy" });
 db.SaveChanges();
 
 /* Update - user < blog */
 Console.WriteLine("Updating the user and adding a blog");
 User user = db.Users.Where(u => u.Name == "Yummy").First();
-db.Blogs.Add(new Blog { Url = "FoodRecipes", User = user });
+db.Blogs.Add(new Blog { Url = "http://foodyummy.blog.com/recipes", User = user });
 db.SaveChanges();
 
 /* Update - user <> post <> blog */
 Console.WriteLine("Updating the blog with a new post and adding it to usern");
-Blog blog = db.Blogs.Where(b => b.Url == "FoodRecipes").First();
+Blog blog = db.Blogs.Where(b => b.Url == "http://foodyummy.blog.com/recipes").First();
 db.Users.First().Posts.Add(new Post { Title = "Taiyaki", Content = "Classic with anko" , User = user, Blog = blog});
 db.SaveChanges();
 
@@ -29,18 +29,18 @@ db.SaveChanges();
 
 //----- Another way to populate data -----//
 /*Create*/
-Console.WriteLine("Inserting - user");
+Console.WriteLine("Inserting > user");
 var user2 = new User
 { 
-    Name = "foodlover" 
+    Name = "foodLover" 
 };
 db.Users.Add(user2);
 db.SaveChanges();
 
-Console.WriteLine("Inserting - category");
+Console.WriteLine("Inserting > category");
 var category2 = new Category
 {
-    Name = "food",
+    Name = "Japan",
 };
 db.Categories.Add(category2);
 db.SaveChanges();
@@ -73,17 +73,17 @@ db.Posts.Add(post2);
 db.SaveChanges();
 
 /*Read*/
-Console.WriteLine("Querying - user");
+Console.WriteLine("Querying >> user");
 user2 = db.Users
-    .Where(u => u.Name == "foodlover")
+    .Where(u => u.Name == "foodLover")
     .FirstOrDefault();
 if (user2 == null)
 {
-    Console.WriteLine("User 'user' not found.");
+    Console.WriteLine("User 'name' not found.");
     return;
 }
 
-Console.WriteLine("Querying - blog");
+Console.WriteLine("Querying >> blog");
 blog2 = db.Blogs
     .Where(b => b.Url == "http://realfoodie.blog.com/reviews")
     .FirstOrDefault();
@@ -93,27 +93,34 @@ if (blog2 == null)
     return;
 }
 
-Console.WriteLine("Querying - post");
+Console.WriteLine("Querying >> post");
+post2 = db.Posts
+.Where(p => p.Title == "Okonomiyaki in Osaka")
+.FirstOrDefault();
 if (post2 == null)
 {
     Console.WriteLine("User 'post' not found.");
     return;
 }
-else
+
+Console.WriteLine("Querying >> category");
+category2 = db.Categories
+.Where(p => p.Name == "Japan")
+.FirstOrDefault();
+if (category2 == null)
 {
-    post2 = db.Posts
-    .Where(p => p.Title == "Okonomiyaki in Osaka")
-    .FirstOrDefault();
-    Console.WriteLine($"{blog2.Url},{post2?.PostId.ToString()},{post2?.Title},{post2?.Content}");
+    Console.WriteLine("Post 'category' not found.");
+    return;
 }
 
+var getPrinter = new DataPrinter(db);
+getPrinter.PrintAllData();
+
 /*Delete*/
-Console.WriteLine("Delete - user");
+Console.WriteLine("\nDelete > user & relating data");
 //db.Remove(user2); ////remove a specific user
 db.Users.RemoveRange(db.Users);
-Console.WriteLine("Delete - post");
-db.Posts.RemoveRange(db.Posts);
-Console.WriteLine("Delete - category");
+Console.WriteLine("Delete > category");
 db.Categories.RemoveRange(db.Categories);
 //db.Remove(category2); ////remove a specific category
 db.SaveChanges();
